seperated method from component class to a new class, removed extra method(update visualization)
